Move 1 func to `theme.py`: get_text_box_theme
