+++ scripts/screens/AllegiancesScreen.py
@@ -5,9 +5,7 @@ import pygame_gui
 from scripts.cat.cats import Cat
 from scripts.game_structure import game
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import event_text_adjust, adjust_list_text
 from ..ui.scale import ui_scale, ui_scale_offset
 from ..clan_package.get_clan_cats import get_alive_clan_queens
