+++ scripts/screens/ChangeGenderScreen.py
@@ -16,9 +16,7 @@ from scripts.game_structure.ui_elements import (
     CatButton,
     UISurfaceImageButton,
 )
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..cat import pronouns
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset, ui_scale_value
