+++ scripts/screens/ChooseMateScreen.py
@@ -11,9 +11,7 @@ from scripts.game_structure.ui_elements import (
     UISpriteButton,
     UISurfaceImageButton,
 )
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..events_module.event_filters import get_personality_compatibility
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
