+++ scripts/screens/ClearingScreen.py
@@ -15,9 +15,7 @@ from scripts.game_structure.ui_elements import (
     UIModifiedImage,
     UIModifiedScrollingContainer,
 )
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..ui.scale import ui_scale
 from .Screens import Screens
