+++ scripts/screens/EventsScreen.py
@@ -32,8 +32,8 @@ from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.icon import Icon
 from scripts.utility import (
     clan_symbol_sprite,
-    get_text_box_theme,
 )
+from scripts.ui.theme import get_text_box_theme
 from scripts.events_module.text_adjust import shorten_text_to_fit
 from scripts.ui.scale import (
     ui_scale,
