+++ scripts/screens/MedDenScreen.py
@@ -14,9 +14,7 @@ from scripts.game_structure.ui_elements import (
     UISurfaceImageButton,
     UIModifiedImage,
 )
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import event_text_adjust, shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_offset
 from ..clan_package.get_clan_cats import find_alive_cats_with_rank
