+++ scripts/screens/MediationScreen.py
@@ -13,9 +13,7 @@ from scripts.game_structure.ui_elements import (
     UISurfaceImageButton,
     UIRelationDisplay,
 )
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_dimensions
 from .Screens import Screens
