+++ scripts/screens/RoleScreen.py
@@ -12,9 +12,7 @@ from scripts.game_structure.ui_elements import (
     UITextBoxTweaked,
     UISurfaceImageButton,
 )
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import adjust_list_text, shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_dimensions
 from .Screens import Screens
