+++ scripts/screens/SwitchClanScreen.py
@@ -11,9 +11,7 @@ from scripts.clan import Clan
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import UIImageButton, UISurfaceImageButton
 from scripts.ui.windows.delete_check import DeleteCheck
-from scripts.utility import (
-    get_text_box_theme,
-)
+from ..ui.theme import get_text_box_theme
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset, ui_scale_value
 from .Screens import Screens
 from .enums import GameScreen
